# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Keops Studios API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Game {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** e.g. Mobile, PC, Mobile & PC */
  platform: string;
  /** @nullable */
  genre?: string | null;
  /** e.g. In Development, Released, Coming Soon */
  status: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  storeUrl?: string | null;
  order: number;
  createdAt?: string;
}