# lib/api-zod/src/generated/types/gameUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Keops Studios API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GameUpdate {
  title?: string;
  description?: string;
  platform?: string;
  genre?: string;
  status?: string;
  imageUrl?: string;
  storeUrl?: string;
  order?: number;
}