# lib/api-zod/src/generated/types/siteContent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Keops Studios API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteContent {
  id: number;
  /** @nullable */
  heroTitleTr?: string | null;
  /** @nullable */
  heroTitleEn?: string | null;
  /** @nullable */
  heroSubtitleTr?: string | null;
  /** @nullable */
  heroSubtitleEn?: string | null;
  /** @nullable */
  aboutTr?: string | null;
  /** @nullable */
  aboutEn?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  discordUrl?: string | null;
  /** @nullable */
  email?: string | null;
}