# lib/api-zod/src/generated/types/siteContentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Keops Studios API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteContentUpdate {
  heroTitleTr?: string;
  heroTitleEn?: string;
  heroSubtitleTr?: string;
  heroSubtitleEn?: string;
  aboutTr?: string;
  aboutEn?: string;
  instagramUrl?: string;
  twitterUrl?: string;
  linkedinUrl?: string;
  youtubeUrl?: string;
  discordUrl?: string;
  email?: string;
}