# lib/api-zod/src/generated/types/studioStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Keops Studios API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudioStats {
  gameCount: number;
  teamSize: number;
  yearsActive: number;
  platforms: string[];
}