# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Keops Studios API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  order: number;
}