# lib/api-zod/src/generated/types/teamMemberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Keops Studios API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMemberUpdate {
  name?: string;
  role?: string;
  bio?: string;
  imageUrl?: string;
  linkedinUrl?: string;
  twitterUrl?: string;
  order?: number;
}